+++ changelog.md
@@ -59,6 +59,7 @@ Performance improvements:
 - `DrawingGroup` cancels drawing when it knows it will not be visible (e.g. outside of current viewport or scissor).
 - Improved caching for `wrappingText:CoordinateToCharacterDisplayIndex(x, y)`
 - Significantly optimised `wrappingText:DisplayIndexToRawIndex(displayIndex)`. On an example file (1229 lines of code), test execution improved from 1.5s to 0.65s - a 65% reduction (including test overhead).
+- `MouseMove` is now buffered to receive only one call per draw frame. If you need more frequent updates, set `dragListener.doNotBufferMouseMove` to true.
 
 Tests:
 - Introduced a test for benchmarking `WrappingText` methods.